change groupId and package names to match convention for releasing artifact to maven central
