--- src/test/java/uk/me/g4dpz/satellite/TestingUtil.java
@@ -1,10 +1,0 @@
-package uk.me.g4dpz.satellite;
-
-public class TestingUtil {
-
-	public static boolean eq(Position position, double lat, double lon,
-			double precision) {
-		return Math.abs(position.getLat() - lat) <= precision
-				&& Math.abs(position.getLon() - lon) <= precision;
-	}
-}
